cmd: move archive cmds to own file
